requirements.txt
amqp==5.3.1
APScheduler==3.11.0
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.7
django-allauth==65.5.0
django-apscheduler==0.7.0
django-debug-toolbar==5.1.0
kombu==5.5.1
packaging==24.2
pillow==11.1.0
pip-review==1.3.0
prompt_toolkit==3.0.50
python-dateutil==2.9.0.post0
redis==5.2.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.2.13
